# api/docs/swagger.yaml
components:
  schemas:
...
    github_com_replicatedhq_kotskinds_multitype.BoolOrString:
      type: object
    types.APIError:
      properties:
        errors:
          items:
            $ref: '#/components/schemas/types.APIError'
          type: array
          uniqueItems: false
        field:
          type: string
        message:
          type: string
        status_code:
          type: integer
      type: object
    types.AppConfig:
      properties:
        groups:
          items:
            $ref: '#/components/schemas/v1beta1.ConfigGroup'
          type: array
          uniqueItems: false
      type: object
    types.AppConfigValue:
      properties:
        data:
          type: string
        dataPlaintext:
          type: string
        default:
          type: string
        filename:
          type: string
        repeatableItem:
          type: string
        value:
          type: string
        valuePlaintext:
          type: string
      type: object
    types.AppConfigValues:
      additionalProperties:
        $ref: '#/components/schemas/types.AppConfigValue'
      type: object
    types.AppConfigValuesResponse:
      properties:
        values:
          $ref: '#/components/schemas/types.AppConfigValues'
      type: object
...
    types.AppUpgrade:
      properties:
        logs:
          type: string
        status:
          $ref: '#/components/schemas/types.Status'
      type: object
...
    types.PatchAppConfigValuesRequest:
      properties:
        values:
          $ref: '#/components/schemas/types.AppConfigValues'
      type: object
...
    types.State:
      enum:
      - Pending
      - Running
      - Succeeded
      - Failed
      example: Succeeded
      type: string
      x-enum-varnames:
      - StatePending
      - StateRunning
      - StateSucceeded
      - StateFailed
    types.Status:
      properties:
        description:
          type: string
        lastUpdated:
          type: string
        state:
          $ref: '#/components/schemas/types.State'
      type: object
    types.TemplateAppConfigRequest:
      properties:
        values:
          $ref: '#/components/schemas/types.AppConfigValues'
      type: object
    types.UpgradeAppRequest:
      properties:
        ignoreAppPreflights:
          type: boolean
      type: object
    v1beta1.ConfigChildItem:
      properties:
        default:
          $ref: '#/components/schemas/github_com_replicatedhq_kotskinds_multitype.BoolOrString'
        name:
          type: string
        recommended:
          type: boolean
        title:
          type: string
        value:
          $ref: '#/components/schemas/github_com_replicatedhq_kotskinds_multitype.BoolOrString'
      type: object
    v1beta1.ConfigGroup:
      properties:
        description:
          type: string
        items:
          items:
            $ref: '#/components/schemas/v1beta1.ConfigItem'
          type: array
          uniqueItems: false
        name:
          type: string
        title:
          type: string
        when:
          type: string
      type: object
    v1beta1.ConfigItem:
      properties:
        affix:
          type: string
        countByGroup:
          additionalProperties:
            type: integer
          type: object
        data:
          type: string
        default:
          $ref: '#/components/schemas/github_com_replicatedhq_kotskinds_multitype.BoolOrString'
        error:
          type: string
        filename:
          type: string
        help_text:
          type: string
        hidden:
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/v1beta1.ConfigChildItem'
          type: array
          uniqueItems: false
        minimumCount:
          type: integer
        multi_value:
          items:
            type: string
          type: array
          uniqueItems: false
        multiple:
          type: boolean
        name:
          type: string
        readonly:
          type: boolean
        recommended:
          type: boolean
        repeatable:
          type: boolean
        required:
          type: boolean
        templates:
          items:
            $ref: '#/components/schemas/v1beta1.RepeatTemplate'
          type: array
          uniqueItems: false
        title:
          type: string
        type:
          type: string
        validation:
          $ref: '#/components/schemas/v1beta1.ConfigItemValidation'
        value:
          $ref: '#/components/schemas/github_com_replicatedhq_kotskinds_multitype.BoolOrString'
        valuesByGroup:
          $ref: '#/components/schemas/v1beta1.ValuesByGroup'
        when:
          type: string
        write_once:
          type: boolean
      type: object
    v1beta1.ConfigItemValidation:
      properties:
        regex:
          $ref: '#/components/schemas/v1beta1.RegexValidator'
      type: object
    v1beta1.GroupValues:
      additionalProperties:
        type: string
      type: object
    v1beta1.RegexValidator:
      properties:
        message:
          type: string
        pattern:
          type: string
      type: object
    v1beta1.RepeatTemplate:
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        name:
          type: string
        namespace:
          type: string
        yamlPath:
          type: string
      type: object
    v1beta1.ValuesByGroup:
      additionalProperties:
        $ref: '#/components/schemas/v1beta1.GroupValues'
      type: object
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  contact:
    email: support@replicated.com
    name: API Support
    url: https://github.com/replicatedhq/embedded-cluster/issues
  description: This is the API for the Embedded Cluster project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Embedded Cluster API
  version: "0.1"
openapi: 3.1.0
paths:
...
  /kubernetes/upgrade/app/config/template:
    post:
      description: Template the app configuration with values for upgrade
      operationId: postKubernetesUpgradeAppConfigTemplate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.TemplateAppConfigRequest'
        description: Template App Config Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.AppConfig'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Template app config for upgrade
      tags:
      - kubernetes-upgrade
  /kubernetes/upgrade/app/config/values:
    get:
      description: Get the current app config values for upgrade
      operationId: getKubernetesUpgradeAppConfigValues
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.AppConfigValuesResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Get the app config values for upgrade
      tags:
      - kubernetes-upgrade
    patch:
      description: Set the app config values with partial updates for upgrade
      operationId: patchKubernetesUpgradeAppConfigValues
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.PatchAppConfigValuesRequest'
        description: Patch App Config Values Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.AppConfigValuesResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Set the app config values for upgrade
      tags:
      - kubernetes-upgrade
  /kubernetes/upgrade/app/status:
    get:
      description: Get the current status of app upgrade
      operationId: getKubernetesUpgradeAppStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.AppUpgrade'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Get app upgrade status
      tags:
      - kubernetes-upgrade
  /kubernetes/upgrade/app/upgrade:
    post:
      description: Upgrade the app using current configuration
      operationId: postKubernetesUpgradeApp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.UpgradeAppRequest'
        description: Upgrade App Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.AppUpgrade'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Upgrade the app
      tags:
      - kubernetes-upgrade
...
servers:
- url: /api